Refactor repository namespaces and update EntityManager usage in tests
