+++ app/Model/Repository/UserRepository.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 namespace App\Model\Repository;
 
-use App\Core\ORM\BaseRepository;
+use App\Model\Repository\BaseRepository;
 use App\Model\Entity\User;
 use App\Core\ORM\EntityManager;
