+++ tests/EntityManagerTest.php
@@ -9,6 +9,7 @@ use App\Core\ORM\SchemaGenerator;
 use PHPUnit\Framework\Attributes\CoversClass;
 use App\Config\Database;
 use App\Config\EnvLoader;
+use App\Model\Repository\UserRepository;
 
 use App\Model\Entity\User;
 
@@ -27,8 +28,8 @@ final class EntityManagerTest extends TestCase
         $schema = new SchemaGenerator($connection);
         $schema->generateTable(User::class);
         $em = new EntityManager(User::class, $connection);
-        $user = $em->find(1);
-
+        $repo = new UserRepository($em);
+        $user = $repo->find(1);
         $this->assertInstanceOf(User::class, $user);
         $this->assertEquals('Sami', $user->getUsername());
     }
@@ -38,7 +39,8 @@ final class EntityManagerTest extends TestCase
         $schema = new SchemaGenerator($connection);
         $schema->generateTable(User::class);
         $em = new EntityManager(User::class, $connection);
-        $result = $em->find(9999); // 9999 does not exist
+        $repo = new UserRepository($em);
+        $result = $repo->find(9999); // 9999 does not exist
 
         $this->assertNull($result);
     }
